requirements.txt
Django==5.0.2
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
python-dotenv
gunicorn
whitenoise
pillow
django-taggit
django-filter
djangorestframework-simplejwt 

